Decouple replies and posts features - Create custom hooks for reply mutation & deletion - Move isReply utility to replies feature directory
