+++ frontend/src/features/replies/hooks/index.js
@@ -1,0 +1,3 @@
+import useReplyDeletion from "./useReplyDeletion";
+import useReplyMutation from "./useReplyMutation";
+export { useReplyDeletion, useReplyMutation };
